requirements.txt
fastapi
uvicorn
openai
pydantic>=2.0
textract
pytesseract
Pillow
python-multipart